Closes #100: You don't want to know.
Ok, if you really need to know: MSVC uses an optimized, non-standard
representation of pointer-to-member. This can be disabled by a compiler setting
or pragma, but this is too dangerous since it produces invalid code when it is
forgotton in some places. Hence this fix opts to fix it by converting the object
pointer to work correctly. It has been tested with single and multiple
inheritance, but not with virtual. YMMV.

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[x] Bugfix
[ ] New Feature
[x] Cleanup
[ ] Optimization
[ ] Documentation