Async fetch implementation version 1
Not very clean as we must store all the details of the k/v pair
so that when the value is retrieved after the fetch, it can be
copied correctly into the buffer supplied in the original call.
We must know the length of the buffer which is written by iRetrieve
into an int we supply. All this info must be managed.

Need to use extra maps internally which significantly bloat
the code.