Change fetch to use a vector of data instead of a map and tweak the test code.
Getting results that appear to be reasonable for read/write. The fix
to the fetch/getfetched calls means that out of order fetch and read is not currently
possible so some work remains.

async fetch is restricted to string for the moment and will template it
and try to make use of existing get<T> code next.